Make an OpenTKUtils project and move relevant code to it
Also split apart the data3d stuff and rename to make it clearer the relationship

OpenTK versions maintained.  No code changes
