+++ EDDiscovery.sln
@@ -1,7 +1,7 @@
 ﻿
 Microsoft Visual Studio Solution File, Format Version 12.00
 # Visual Studio 15
-VisualStudioVersion = 15.0.26403.7
+VisualStudioVersion = 15.0.27130.2027
 MinimumVisualStudioVersion = 10.0.40219.1
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EDDiscovery", "EDDiscovery\EDDiscovery.csproj", "{6EE10D92-E3E4-44D9-8E6B-851801B0B510}"
 	ProjectSection(ProjectDependencies) = postProject
@@ -48,6 +48,8 @@ EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "TestNetLogEntry", "TestNetLogEntry\TestNetLogEntry.csproj", "{913F9FFA-C90A-4DFE-B613-87157CA48FD9}"
 EndProject
 Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "EDDIcons", "EDDIcons\EDDIcons.csproj", "{646F71BE-F102-444E-9254-818E5A320AB8}"
+EndProject
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "OpenTKUtils", "OpenTKUtils\OpenTKUtils.csproj", "{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}"
 EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
@@ -201,6 +203,18 @@ Global
 		{646F71BE-F102-444E-9254-818E5A320AB8}.Release|x64.Build.0 = Release|Any CPU
 		{646F71BE-F102-444E-9254-818E5A320AB8}.Release|x86.ActiveCfg = Release|Any CPU
 		{646F71BE-F102-444E-9254-818E5A320AB8}.Release|x86.Build.0 = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|x64.Build.0 = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Debug|x86.Build.0 = Debug|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|Any CPU.Build.0 = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|x64.ActiveCfg = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|x64.Build.0 = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|x86.ActiveCfg = Release|Any CPU
+		{4CFA37D5-4182-4CAE-AFC3-7563A13B44EE}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE
