+++ OpenTKUtils/Keyboard/KeyboardActions.cs
@@ -13,14 +13,12 @@
  * 
  * EDDiscovery is not affiliated with Frontier Developments plc.
  */
-using OpenTK.Input;
+
 using System;
 using System.Collections.Generic;
-using System.Linq;
-using System.Text;
 using System.Windows.Forms;
 
-namespace EDDiscovery._3DMap
+namespace OpenTKUtils
 {
     public class KeyboardActions
     {
