+++ OpenTKUtils/packages.config
@@ -1,0 +1,5 @@
+﻿<?xml version="1.0" encoding="utf-8"?>
+<packages>
+  <package id="OpenTK" version="2.0.0" targetFramework="net46" />
+  <package id="OpenTK.GLControl" version="1.1.2349.61993" targetFramework="net46" />
+</packages>
